add screendensity for devinfo

+++ include/uv_ext.h
@@ -112,7 +112,7 @@ struct uv_devinfo_s {
     int screenwidth;
     int screenheight;
     int devicetype;
-    double screendensity;
+    float screendensity;
 };
 
 /****************************************************************************
